Split out Timestep tests into own file
